refactor(rails): follow best practices
- registrations_controller: params.expect with fallback (Rails 8)
- onboarding_controller: i18n t('.name_required') etc
- calendar_events_controller: verify_authorized for create/new, DRY period_ranges, authorize new/create, fix duplicate create method
- owned_record_policy: add create?/new?
- config/routes: remove duplicate calendar/tracking/period, fix duplicate get calendar alias
- appointment_date_picker_controller: fix setSlotTime24 arg, fmtTime24 alias, remove dead _legacyStage, decouple From/To, exitInline
- views: extract _date_picker_modal partial, DRY new/edit (761/715 from 830/784), unify repeatUntil to w-380 flex h-7
- rebuild tailwind.css

+++ app/controllers/calendar_events_controller.rb
@@ -1,6 +1,6 @@
 class CalendarEventsController < ApplicationController
   before_action :set_event, only: [:show, :edit, :update, :destroy]
-  after_action :verify_authorized, only: [:show, :edit, :update, :destroy]
+  after_action :verify_authorized, only: [:show, :edit, :update, :destroy, :create]
 
   def index
     redirect_to forecast_path
@@ -13,6 +13,7 @@ class CalendarEventsController < ApplicationController
     @event = current_user.calendar_events.build(
       date: params[:date] || Time.zone.today
     )
+    authorize @event
     @period_ranges = period_ranges_json
     session[:appointment_return_to] = request.referer if request.referer.present?
   end
@@ -24,6 +25,7 @@ class CalendarEventsController < ApplicationController
 
   def create
     @event = current_user.calendar_events.build(event_params)
+    authorize @event
     if @event.save
       redirect_to session.delete(:appointment_return_to) || calendar_path(date: @event.date, appointment_scheduled: "1"), notice: t(".created")
     else
